server/requirements.txt
paramiko
tweepy
config
python-dotenv
